Fixes linking examples with targets that are not built
CMake built ArrayFireConfig files for each of the backend when they are built.
If you disabled a backend after it was built, the config file created a target
for the backend even though it doesn't exist. This commit modifys the config
file so that it only creates the target if the target's library also exists.

The side effect of this change is that if the library exists but the
BUILD_<backend> is off then it will build the examples anyway.